Move Xcode build settings into configuration files.

+++ ZeroKitURLConnectionManager.h
@@ -17,7 +17,7 @@
 
 - (NSArray *)activeConnectionIdentifiers;
 
-- (int)numberOfActiveConnections;
+- (NSInteger)numberOfActiveConnections;
 
 #pragma mark -
